requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
beanie==1.25.0
python-dotenv==1.0.0
httpx==0.26.0
loguru==0.7.2
pytest==7.4.4
pytest-asyncio==0.23.5
pytest-mock==3.12.0
asgi-lifespan==2.1.0
pytest-cov==4.1.0
pytest-env==1.0.1
motor==3.3.2
pymongo==4.6.1 